evals/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "skill-eval"
version = "0.1.0"
description = "Evaluation harness for Speakeasy agent skills"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.40.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "click>=8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[project.scripts]
skill-eval = "skill_eval.__main__:cli"

[tool.hatch.build.targets.wheel]
packages = ["skill_eval"]
